# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * App-managed role, read-only. Assigned in the database.
 * @nullable
 */
export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole] | null;


export const AuthUserRole = {
  admin: 'admin',
  pm: 'pm',
  finance: 'finance',
  coordinator: 'coordinator',
} as const;

export type AuthUserTheme = typeof AuthUserTheme[keyof typeof AuthUserTheme];


export const AuthUserTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  /**
     * App-managed role, read-only. Assigned in the database.
     * @nullable
     */
  role: AuthUserRole;
  /**
     * User-chosen display name, separate from the synced first/last name.
     * @nullable
     */
  displayName: string | null;
  /**
     * Effective avatar URL to display (uploaded photo if present, otherwise the synced profile image).
     * @nullable
     */
  avatarUrl: string | null;
  /** True when the user has uploaded a personal photo. */
  hasCustomAvatar: boolean;
  theme: AuthUserTheme;
  language: string;
  currency: string;
}

export type UpdateProfileRequestTheme = typeof UpdateProfileRequestTheme[keyof typeof UpdateProfileRequestTheme];


export const UpdateProfileRequestTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface UpdateProfileRequest {
  /**
     * @maxLength 120
     * @nullable
     */
  displayName?: string | null;
  theme?: UpdateProfileRequestTheme;
  /**
     * @minLength 2
     * @maxLength 10
     */
  language?: string;
  /**
     * @minLength 3
     * @maxLength 3
     */
  currency?: string;
}

export interface SetAvatarRequest {
  /**
     * The uploaded object path (e.g. `/objects/uploads/uuid`) or full GCS URL returned by the upload flow.
     * @minLength 1
     */
  avatarUrl: string;
}

export interface UploadUrlRequest {
  /**
     * Original file name.
     * @minLength 1
     */
  name: string;
  /**
     * File size in bytes.
     * @minimum 1
     */
  size: number;
  /**
     * MIME type of the file (e.g. `image/jpeg`).
     * @minLength 1
     */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};