# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from './authUserRole';
import type { AuthUserTheme } from './authUserTheme';

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  /**
     * App-managed role, read-only. Assigned in the database.
     * @nullable
     */
  role: AuthUserRole;
  /**
     * User-chosen display name, separate from the synced first/last name.
     * @nullable
     */
  displayName: string | null;
  /**
     * Effective avatar URL to display (uploaded photo if present, otherwise the synced profile image).
     * @nullable
     */
  avatarUrl: string | null;
  /** True when the user has uploaded a personal photo. */
  hasCustomAvatar: boolean;
  theme: AuthUserTheme;
  language: string;
  currency: string;
}